+++ .gitignore
@@ -3,3 +3,5 @@ vendor
 .env
 test.sql
 test.sqlite
+
+main.go
