+++ go.mod
@@ -3,6 +3,7 @@ module github.com/sharovik/orm
 go 1.16
 
 require (
+	github.com/go-sql-driver/mysql v1.6.0 // indirect
 	github.com/mattn/go-sqlite3 v1.14.6
 	github.com/stretchr/testify v1.7.0
 )
